+++ Changelog.rst
@@ -1,8 +1,10 @@
 Version NEXTVERSION
-----------------
+--------------
 
 **2026-06-??**
 
+* New keyword parameter to `cf.Field.regrids` and `cf.Field.regridc`:
+  ``mtol`` (https://github.com/NCAS-CMS/cf-python/issues/949)
 * Fix bug in `cf.read` that prevented some OPeNDAP URLS being read
   (https://github.com/NCAS-CMS/cf-python/issues/948)
 * Fix bug that caused weighted `cf.collapse` and `cf.weights` to fail